refactor: 로직과 ui 분리
