+++ src/main/java/baseball/ui/GameStatus.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.ui;
 
 public enum GameStatus {
     GAME_START, GAME_END
