+++ src/test/java/baseball/domain/BaseballRefereeTest.java
@@ -2,6 +2,7 @@ package baseball.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.ui.GameResult;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
